pyproject.toml
[tool.poetry]
name = "tasks"
version = "0.1.0"
description = ""
authors = ["ramdirt <ramdirt@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
redis = "^5.0.4"
python-jose = "^3.3.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
